# perf: Increase performance to bind a list of GTS