--- a/pyproject.toml
+++ b/pyproject.toml
@@ -11,12 +11,13 @@ pandas = "^1.2.4"
 pyarrow = "^4.0.1"
 dateparser = "^1.0.0"
 durations = "^0.3.3"
+tqdm = "^4.62.3"
 
 [tool.poetry.dev-dependencies]
 black = "^21.5b2"
 pytest = "^6.2.4"
 coverage-badge = "^1.0.1"
-coverage = "^5.5"
+coverage = "^6.1.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]