--- a/test_gts.py
+++ b/test_gts.py
@@ -1,6 +1,38 @@
-from gts import LGTS
-import pytest
+from gts import LGTS, GTS, is_lgts, is_gts
 import pandas as pd
+import pytest
+
+
+gts = {
+    "timestamps": range(10),
+    "values": range(10),
+    "classname": "metric",
+    "labels": {"foo": "bar"},
+}
+
+
+def test_gts():
+    empty_gts = gts.copy()
+    empty_gts["timestamps"] = []
+    empty_gts["values"] = []
+    res = GTS(empty_gts)
+    pd.testing.assert_frame_equal(
+        res.data, pd.DataFrame({"classname": "metric", "foo": "bar"}, index=[0])
+    )
+    missing_classname_gts = gts.copy()
+    del missing_classname_gts["classname"]
+    assert not is_gts(missing_classname_gts)
+    toomuch_columns_gts = gts.copy()
+    toomuch_columns_gts["foo"] = "bar"
+    assert not is_gts(toomuch_columns_gts)
+
+
+def test_lgts():
+    lgts = [gts]
+    assert is_lgts(lgts)
+    assert len(LGTS(lgts)) == 10
+    assert not is_lgts([])
+    assert not is_lgts([1])
 
 
 def test_dataframe():