--- a/tests/test_pywarp10.py
+++ b/tests/test_pywarp10.py
@@ -1,4 +1,4 @@
-from pywarp10 import SanitizeError, Warpscript
+from pywarp10.pywarp10 import SanitizeError, Warpscript
 import pytest
 import pandas as pd