Binding a list of GTS could take a long time. This change allow to create a panda DataFrame much faster.
For 200 GTS to bind together (each with 1000 values and 5 non intersecting labels), we go from 20 s to 700 ms which correspond to a gain of 96.5%.

# Follow up benchmark

## Old implementation

2021-11-26 15:00:51,627 Time to generate 1 gts: 9 ms
2021-11-26 15:00:51,659 Time to generate 5 gts: 31 ms
2021-11-26 15:00:51,770 Time to generate 10 gts: 106 ms
2021-11-26 15:00:52,046 Time to generate 15 gts: 270 ms
2021-11-26 15:00:52,725 Time to generate 20 gts: 670 ms
2021-11-26 15:00:54,135 Time to generate 25 gts: 1398 ms
2021-11-26 15:00:57,244 Time to generate 50 gts: 3088 ms
2021-11-26 15:01:04,502 Time to generate 100 gts: 7215 ms
2021-11-26 15:01:24,500 Time to generate 200 gts: 19919 ms

## List of Dict --> List of GTS object

### Old implementation

2021-11-26 15:03:44,104 Time to generate 1 gts: 2 ms
2021-11-26 15:03:44,112 Time to generate 5 gts: 6 ms
2021-11-26 15:03:44,129 Time to generate 10 gts: 12 ms
2021-11-26 15:03:44,152 Time to generate 15 gts: 15 ms
2021-11-26 15:03:44,181 Time to generate 20 gts: 20 ms
2021-11-26 15:03:44,216 Time to generate 25 gts: 23 ms
2021-11-26 15:03:44,282 Time to generate 50 gts: 44 ms
2021-11-26 15:03:44,407 Time to generate 100 gts: 83 ms
2021-11-26 15:03:44,649 Time to generate 200 gts: 163 ms
2021-11-26 15:03:45,245 Time to generate 500 gts: 401 ms
2021-11-26 15:03:46,461 Time to generate 1000 gts: 819 ms

### Initialize list first

2021-11-26 15:09:12,653 Time to generate 1 gts: 3 ms
2021-11-26 15:09:12,662 Time to generate 5 gts: 6 ms
2021-11-26 15:09:12,682 Time to generate 10 gts: 13 ms
2021-11-26 15:09:12,707 Time to generate 15 gts: 16 ms
2021-11-26 15:09:12,736 Time to generate 20 gts: 20 ms
2021-11-26 15:09:12,769 Time to generate 25 gts: 23 ms
2021-11-26 15:09:12,838 Time to generate 50 gts: 46 ms
2021-11-26 15:09:12,967 Time to generate 100 gts: 86 ms
2021-11-26 15:09:13,207 Time to generate 200 gts: 162 ms
2021-11-26 15:09:13,841 Time to generate 500 gts: 432 ms
2021-11-26 15:09:15,074 Time to generate 1000 gts: 834 ms

## GTS contains a distinct column

2021-11-26 16:58:16,724 Time to generate 1 gts: 7 ms
2021-11-26 16:58:16,742 Time to generate 5 gts: 18 ms
2021-11-26 16:58:16,783 Time to generate 10 gts: 40 ms
2021-11-26 16:58:16,847 Time to generate 15 gts: 62 ms
2021-11-26 16:58:16,939 Time to generate 20 gts: 90 ms
2021-11-26 16:58:17,062 Time to generate 25 gts: 120 ms
2021-11-26 16:58:17,271 Time to generate 50 gts: 206 ms
2021-11-26 16:58:17,666 Time to generate 100 gts: 386 ms
2021-11-26 16:58:18,403 Time to generate 200 gts: 721 ms